anti/requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
pypdf
requests
python-multipart
numpy
